move constraints to separate module
